Only implement countNodes once
